backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
celery[redis]>=5.6,<5.7
django-celery-beat>=2.8,<2.10
psycopg[binary]>=3.2,<3.3
redis>=6.2,<6.3
uvicorn[standard]>=0.38,<0.39
whitenoise>=6.11,<6.12
pandas>=2.2,<3.1
yfinance>=0.2.54,<1.3
curl-cffi>=0.7.4,<0.15
